+++ api/v1/endpoints/playbooks.py
@@ -1,29 +1,28 @@
-from fastapi import APIRouter, HTTPException
-import yaml
 import os
+import yaml
 import glob
 import re
-import subprocess
+from pathlib import Path
+from fastapi import APIRouter
+
 
 router = APIRouter()
 
-# Playbooks path
-base_path = "/ansible-playbooks/playbooks/"
-scripts_path = "/ansible-playbooks/scripts/"
-crontab_path = '/etc/cron.d/ansible'
+playbooks_path = Path(os.getenv('ANSIBLE_PLAYBOOKS'))
+crontab_path = Path('/etc/cron.d/ansible')
 
-@router.get("/get-playbooks")
+@router.get("/")
 def get_playbooks():
     """
     This endpoint reads all .yaml files in the specified path and checks if it is scheduled in crontab.
     Each playbook is added to the 'playbooks' list.
     """
     # Check if path exists
-    if not os.path.exists(base_path):
+    if not os.path.exists(playbooks_path):
         return {"error": "The specified path does not exist."}
 
     # Fetch every playbook
-    yaml_files = glob.glob(os.path.join(base_path, '*.yaml'))
+    yaml_files = glob.glob(os.path.join(playbooks_path, '/', '*.yaml'))
 
     playbooks = {}
 
@@ -70,21 +69,3 @@ def get_playbooks():
                 return {"error": f"Error reading {yaml_file}."}
 
     return playbooks
-
-@router.post("/execute-script/{script_name}")
-def execute_script(script_name: str):
-    """
-    This endpoint executes a script located in /ansible-playbooks/scripts/.
-    The script name is passed as a path parameter.
-    """
-    script_path = os.path.join(scripts_path, script_name)
-
-    # Check if script exists
-    if not os.path.exists(script_path):
-        raise HTTPException(status_code=404, detail="Script not found")
-
-    # Execute the script
-    result = subprocess.run(["bash", script_path], capture_output=True, text=True)
-
-    # Return the result
-    return {"output": result.stdout, "error": result.stderr}
